+++ src/AppHost/appsettings.json
@@ -5,5 +5,9 @@
       "Microsoft.AspNetCore": "Warning",
       "Aspire.Hosting.Dcp": "Warning"
     }
+  },
+  "Parameters":{
+    "CustomDomain": "https://c5m.ca",
+    "DefaultRedirectUrl": "https://azure.com"
   }
 }
