+++ src/FunctionsLight/Cloud5mins.ShortenerTools.FunctionsLight.csproj
@@ -12,22 +12,13 @@
     <!-- <PackageReference Include="Microsoft.ApplicationInsights.WorkerService" Version="2.22.0" /> -->
     <!-- <PackageReference Include="Microsoft.Azure.Functions.Worker.ApplicationInsights" Version="2.0.0" /> -->
     <PackageReference Include="Microsoft.Azure.Functions.Worker" Version="2.0.0" />
-<<<<<<<< HEAD:src/Cloud5mins.ShortenerTools.FunctionsLight/Cloud5mins.ShortenerTools.FunctionsLight.csproj
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="2.0.0" />
-========
     <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="2.0.2" />
->>>>>>>> main:src/FunctionsLight/Cloud5mins.ShortenerTools.FunctionsLight.csproj
     <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.3.0" />
     <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http.AspNetCore" Version="2.0.1" />
   </ItemGroup>
   <ItemGroup>
-<<<<<<<< HEAD:src/Cloud5mins.ShortenerTools.FunctionsLight/Cloud5mins.ShortenerTools.FunctionsLight.csproj
-    <ProjectReference Include="..\Cloud5mins.ShortenerTools.ServiceDefaults\Cloud5mins.ShortenerTools.ServiceDefaults.csproj" />
-    <ProjectReference Include="..\Cloud5mins.ShortenerTools.Core\Cloud5mins.ShortenerTools.Core.csproj" />
-========
     <ProjectReference Include="..\ServiceDefaults\Cloud5mins.ShortenerTools.ServiceDefaults.csproj" />
     <ProjectReference Include="..\Core\Cloud5mins.ShortenerTools.Core.csproj" />
->>>>>>>> main:src/FunctionsLight/Cloud5mins.ShortenerTools.FunctionsLight.csproj
   </ItemGroup>
   <ItemGroup>
     <None Update="host.json">
