--- src/TinyBlazorAdmin/Components/Components/Layout/MainLayout.razor
@@ -1,39 +1,0 @@
-﻿@inherits LayoutComponentBase
-
-<FluentLayout>
-    <FluentHeader>
-        Tiny Blazor Admin
-    </FluentHeader>
-    <FluentStack Class="main" Orientation="Orientation.Horizontal" Width="100%">
-        <NavMenu />
-        <FluentBodyContent Class="body-content">
-            <div class="content">
-                @Body
-            </div>
-        </FluentBodyContent>
-    </FluentStack>
-    <FluentFooter>
-        <div class="link1">
-            <a href="https://www.fluentui-blazor.net" target="_blank">Documentation and demos</a>
-        </div>
-        <div class="link2">
-            <a href="https://learn.microsoft.com/en-us/aspnet/core/blazor" target="_blank">About Blazor</a>
-        </div>
-    </FluentFooter>
-</FluentLayout>
-
-<FluentDialogProvider @rendermode="RenderMode.InteractiveServer" />
-<FluentToastProvider @rendermode="RenderMode.InteractiveServer" />
-<FluentTooltipProvider @rendermode="RenderMode.InteractiveServer" />
-
-<div id="blazor-error-ui">
-    An unhandled error has occurred.
-    <a href="" class="reload">Reload</a>
-    <a class="dismiss">🗙</a>
-</div>
-
-
-@code {
-    public DesignThemeModes Mode { get; set; }
-    public OfficeColor? OfficeColor { get; set; }
-}
