--- src/TinyBlazorAdmin/Components/Components/Layout/NavMenu.razor
@@ -1,21 +1,0 @@
-﻿@rendermode InteractiveServer
-
-<div class="navmenu">
-    <input type="checkbox" title="Menu expand/collapse toggle" id="navmenu-toggle" class="navmenu-icon" />
-    <label for="navmenu-toggle" class="navmenu-icon"><FluentIcon Value="@(new Icons.Regular.Size20.Navigation())" Color="Color.Fill" /></label>
-    <nav class="sitenav" aria-labelledby="main-menu" onclick="document.getElementById('navmenu-toggle').click();">
-        <FluentNavMenu Id="main-menu" Collapsible="true" Width="250" Title="Navigation menu" @bind-Expanded="expanded">
-            <FluentNavLink Href="/" Match="NavLinkMatch.All" Icon="@(new Icons.Regular.Size20.Home())" IconColor="Color.Accent">Home</FluentNavLink>
-            <FluentNavLink Href="urlmanager" Icon="@(new Icons.Regular.Size20.AppsListDetail())" IconColor="Color.Accent">URL Manager</FluentNavLink>
-            <FluentNavLink Href="/statistics" Match="NavLinkMatch.All" Icon="@(new Icons.Regular.Size20.ChartMultiple())">Statistics</FluentNavLink>
-            <FluentNavLink Href="/settings" Match="NavLinkMatch.All" Icon="@(new Icons.Regular.Size20.TextBulletListSquareSettings())">Settings</FluentNavLink>
-            <FluentNavLink Href="/help" Match="NavLinkMatch.All" Icon="@(new Icons.Regular.Size20.QuestionCircle())">Help</FluentNavLink>
-        </FluentNavMenu>
-    </nav>
-</div>
-
-@code {
-    private bool expanded = true;
-
-}
-
