--- src/TinyBlazorAdmin/Components/Components/Pages/Help.razor
@@ -1,11 +1,0 @@
-@page "/help"
-
-<PageTitle>Tiny Blazor Admin - Help</PageTitle>
-
-<h1>Help</h1>
-<FluentCard Width="50%" Height="50%">
-    <ul>
-        <li>You need to be authenticated to have access.  By default Active Directory is the one used (any outlook.com). Others (GitHub, Twitter, etc.) are also available to used.</li>
-        <li>You need to be part or the role <b>admin</b> (all lowercase) to have access. This is manage from the Azure Portal, in the <i>Role management</i> blade of your Azure Static Web App. Refer to the <a href="https://github.com/FBoucher/TinyBlazorAdmin">documentation</a> for more details.</li>
-    </ul>
-</FluentCard>
