--- src/TinyBlazorAdmin/Components/Components/Pages/Home.razor
@@ -1,20 +1,0 @@
-﻿@page "/"
-
-<PageTitle>Tiny Blazor Admin</PageTitle>
-
-<h1>Tiny Blazor Admin</h1>
-<div>
-    <p>
-    Welcome to the <b><a href="https://github.com/FBoucher/TinyBlazorAdmin" target="_blank">Tiny Blazor Admin</a></b>
-    </p>
-
-    <div>
-    <p>This is the frontend of a solution using <a href="https://github.com/FBoucher/AzUrlShortener" target="_blank">AzUrlShortener</a> both projects are available on GitHub.</p>
-    
-    <p>This site is build using <a href="http://blazor.net/" target="_blank">.Net Blazor</a> and some Blazor components from <a href="https://www.fluentui-blazor.net/" target="_blank">FluentUI Blazor</a> and the chart are from <a href="https://demos.blazorbootstrap.com/" target="_blank">BlazorBootstrap</a></p>
-
-        <p>If you want to learn more about Tiny Blazor Admin please refer to the <a href="https://github.com/microsoft/AzUrlShortener/blob/main/doc/faq.md" target="_blank">documentation</a>.</p>
-    </div>
-
-    <div style="align-content: right;"><img class="mainLogo" src="/images/TinyBlazorAdmin.png" /></div>
-</div>
