--- src/TinyBlazorAdmin/Components/Components/_Imports.razor
@@ -1,12 +1,0 @@
-﻿@using System.Net.Http
-@using System.Net.Http.Json
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Routing
-@using Microsoft.AspNetCore.Components.Web
-@using static Microsoft.AspNetCore.Components.Web.RenderMode
-@using Microsoft.AspNetCore.Components.Web.Virtualization
-@using Microsoft.FluentUI.AspNetCore.Components
-@using Microsoft.JSInterop
-@using Cloud5mins.ShortenerTools.TinyBlazorAdmin
-@using Cloud5mins.ShortenerTools.TinyBlazorAdmin.Components
-@using Icons = Microsoft.FluentUI.AspNetCore.Components.Icons
